modification about role  of commmercial

+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\User;
 use Illuminate\Support\Facades\Validator;
+use RealRashid\SweetAlert\Facades\Alert;
 
 class UserController extends Controller
 {
@@ -14,13 +15,20 @@ class UserController extends Controller
     }
 
     public function AccessRight(Request $request){
-  
-
       $user =User::find($request->userid);
+
+      if($user->is_admin==0){
+        $user->is_admin = 4;
+        $user->save();
+        return redirect()->back();
+
+      }
+
       if($request->etat==1){
         $user->is_admin = 1;
         $user->save();
-      }else{
+      }
+      else{
         $user->is_admin = 2;
         $user->save();
       }
@@ -31,14 +39,24 @@ class UserController extends Controller
 
     public function AccessDenied(Request $request){
       $user =User::find($request->userid);
+      if($user->is_admin==4){
+        $user->is_admin = 0;
+      }else{
         $user->is_admin = 3;
-        $user->save();
+      }
+      $user->save();
         return redirect()->back();
 
     }
 
     public function DeleteUser($id){
-      User::destroy($id);
+      $users= User::find($id);
+      if($users->is_admin==4 || $users->is_admin==0){
+         Alert::html('impossible ce supprimer cette utilisateur a partir de cette espece','veillez vous rendre sur l espace commercial','warning');
+      }else{
+        User::destroy($id);
+
+      }
       return redirect()->back();
     }
 }
